backend/requirements.txt
fastapi
uvicorn
pandas
google-api-python-client
google-auth-httplib2
google-auth
